+++ packages/tools/devtools/devtools-view/src/TelemetryUtils.ts
